ARCHBOM-1597: Use a virtualenv for `make upgrade` in generated project (#73)
Without this, the pip commands run in the virtualenv of
edx-cookiecutters itself, which causes a very difficult-to-diagnose
flaky test result on Travis CI, possibly involving Travis's pip cache.

Adds `virtualenv` to cookiecutter's dependencies.

Clearing the environment for the subprocess call is essential --
otherwise, the PATH and VIRTUALENV environment variables of the test
runner will leak through to the subprocess.

The virtualenv is created and destroyed again just for the upgrade
command, since it interferes with other tests. For instance, the
pylint and isort checks will try to lint site-packages in the
virtualenv. To fix this we'll want to call a `make quality` command in
the baked project rather than parastically relying on the packages and
configurations in the cookiecutters project, but that's out of scope
for this fix. See ARCHBOM-1593 for future improvements.