build: Ensure pip.txt in template gets updated.
When the pip.txt file was added, it was not added to the list of
template requirements files that should get updated as a part of the
routine `make upgrade` jobs.  This resulted in pip becoming out of date.

This change updates the make target to upgrade pip.txt along with
pip-tools.txt and also runs `make upgrade_template` to catch pip up to
the latest version.