fix: get codecov upload working & correct .coveragerc
1.

First, make .coveragerc use project_name, not repo_name, for source

Per https://coverage.readthedocs.io/en/6.5.0/config.html#run-source, the
`source` of a .coveragerc [run] section is "A list of packages or
directories, the source to measure during execution.".

The name of the directory with the code is set by
cookiecutter-django-ida to be {{cookiecutter.project_name}}, not
{{cookiecutter.repo_name}}.

Thus, the right value to use for the .coveragerc source is
{{cookiecutter.project_name}}. This commit makes that happen.

2.

Second, change coverage to run on the py38 tox matrix job instead of the
quality matrix job.

The code coverage job uploads a code coverage file generated during the
tests. However, the tests are run in py38, not in quality.

When quality is the target for the coverage job, I get the following
error when running codecov:

```
[2023-04-12T01:27:35.261Z] ['info'] => Project root located at: /home/runner/work/commerce-coordinator/commerce-coordinator
[2023-04-12T01:27:35.265Z] ['info'] -> No token specified or token is empty
[2023-04-12T01:27:35.419Z] ['info'] Searching for coverage files...
[2023-04-12T01:27:35.488Z] ['info'] Warning: Some files located via search were excluded from upload.
[2023-04-12T01:27:35.489Z] ['info'] If Codecov did not locate your files, please review https://docs.codecov.com/docs/supported-report-formats
[2023-04-12T01:27:35.489Z] ['error'] There was an error running the uploader: No coverage files located, please try use `-f`, or change the project root with `-R`
Error: Codecov: Failed to properly upload: The process '/home/runner/work/_actions/codecov/codecov-action/v3/dist/codecov' failed with exit code 255
```

For an example, see:

https://github.com/edx/commerce-coordinator/actions/runs/4673476499/jobs/8277003234