build: remove deprecated codecov CI package in requirements/ci.in
The codecov PyPI package is now deprecated. See:

https://community.codecov.com/t/codecov-yanked-from-pypi-all-versions/4259

Its presence in our requirements files are currently causing a lot of CI
failures.

It is replaced by codecov-action, a GitHub Action that will upload
coverage results to Codecov.

codecov-action is already in edx-cookiecutters in the .workflow folder
of python-template, on which all other cookiecutters currently depend.