fix: Remove manage.py reference from base template; unit-test library tests
The base template's Makefile had a call to `pylint ... manage.py *.py`
which is inappropriate for some of the cookiecutters, as not all of them
create a manage.py. Simply removing the explicit reference is fine, since
it's already covered by the `*.py`.

This also simplifies the `test_upgrade` and `test_quality` tests for the
cookiecutter by running the existing make targets in a virtualenv in the
baked output, rather than reproducing the full list of checks in test code.

(I haven't tackled the other cookiecutters yet, since some of them
actually have test failures if you try to use this approach, or don't yet
have a `make test-all` or equivalent.)