fixup! Prevent duplicate target error; deduplicate readme
The "Getting Help" section creates an implicit duplicate target that
conflicts with the `_Getting Help:` reference. Rather than doing something
more complicated, just inline the link.

I also noticed this "getting started" text was copied into the cookiecutter
readmes, which are *instructions* for using the cookiecutters, not used in
the template output. This duplicates the root readme in the repo, so I
removed that text. (Same for in the repo's own readme.)

(This required using a double-underscore link to make it anonymous,
otherwise you *still* get an implicit target!)

At this point, tests should pass.