feat: Install pkg-config for IDA Dockerfile (#364)
mysqlclient 2.2.0 requires Debian package pkg-config to be installed. In
some Dockerfiles this also leads to tzdata being installed, which prompts
for config and causes the image build to hang, so add
`DEBIAN_FRONTEND=noninteractive` to prevent the prompt.

Also organize packages by general purpose. (Would be nice to have comments
on all of the groups, but might as well start with this one.)

Fixes https://github.com/openedx/edx-cookiecutters/issues/363