fix: Don't filter out extras for a pip dependency.
Before this change, the function was dropping extra dependencies. For
example:

If you had a requirements file like the following:
```
Xblock[django]
```

It will output: `Xblock`

The correct output should be `Xblock[django]`

This drops the extras which may have further constraint requirements or
more packages that are required. In the example, the package would need
Django but that dependency would not be visible upstream correctly.