feat!: Update cookiecutters to Python 3.12
Even thought we were previously running tests on python 3.11, the
cookiecutters themselves were not updated to run on 3.11.  So they
failed when 3.8 was not available on the host system.

Update everything to run on python 3.12 including the defaults for the
cookiecut repos so that any new systems are creted with the latest
version of python we support as a community.