feat: [django-ida] allow logging format to be overridden
As part of ENT-10092 to correlate logs with DataDog traces, we need the
ability to add trace/span IDs to the log lines. However, this being an
open source project, we cannot assume a particular format and
observability tool. Hence the need to allow operators to completely
override the log format.

ENT-10092